server/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
Pillow==9.4.0
requests==2.28.2
urllib3==1.26.15
Werkzeug==2.2.3
